requirements.txt
requests==2.28.2
BeautifulSoup4==4.12.2
python-dotenv==1.0.0
python-telegram-bot
SQLAlchemy==2.0.9
postgres

APScheduler==3.10.1

Jinja2~=3.1.2